Validacion de nombre del rol al modificar

+++ Prototipo/controllers/rol.controller.js
@@ -3,7 +3,7 @@ const Rol = require('../models/rol.model');
 
 exports.get_crearRol = async (request, response, next) => {
     try {
-        const privilegios = await Rol.privilegioAll(); // Suponiendo que tienes un método fetchAll en tu modelo de Privilegio
+        const privilegios = await Rol.privilegioAll();
         response.render('crearRol', {
             nombreRol: request.session.nombreRol || '',
             csrfToken: request.csrfToken(),
@@ -21,7 +21,7 @@ exports.get_crearRol = async (request, response, next) => {
 
 exports.post_crearRol = async (request, response, next) => {
     const nuevo_Rol = new Rol(request.body.nombreRol);
-    request.body.nombreRol
+    
 
     try {
         const rolExiste = await Rol.findByNombre(request.body.nombreRol);
@@ -63,6 +63,7 @@ exports.get_editarRol = async (req, res, next) => {
             res.render("editarRol", {
                 nombreRol: req.session.nombreRol || '',
                 csrfToken: req.csrfToken(),
+                error: req.session.error || '',
                 permisos: req.session.permisos || [],
                 rol: roles[0], // Asegúrate de que rol tenga la estructura esperada, como { id: ..., nombre: ... }
                 privilegios: privilegios,
@@ -80,32 +81,41 @@ exports.get_editarRol = async (req, res, next) => {
 
 exports.post_editarRol = async (req, res, next) => {
     try {
-        // Actualizar el nombre del rol en la base de datos
-        console.log("Current body = ", req.body)
-        const idRol = req.body.idRol;
-        const nuevoNombreRol = req.body.nombreRol;
-        await Rol.update(idRol, nuevoNombreRol);
+        const rolExistente = await Rol.findByNombre(req.body.nombreRol);
 
-        // Verificar si hay privilegios seleccionados
-        if (req.body.rol) {
-            console.log([req.body.rol].flat())
-            // Eliminar todos los privilegios asociados al rol
-            await Rol.eliminarPrivilegios(idRol);
+        if (rolExistente) {
+            req.session.error = 'Rol ya existente. Por favor, coloca otro nuevo.';
+            res.redirect(`/editarRol/${req.body.nombreRol}`);
+        } else {
+            req.session.error = ''; 
+            // Actualizar el nombre del rol en la base de datos
+            console.log("Current body = ", req.body)
+            const idRol = req.body.idRol;
+            const nuevoNombreRol = req.body.nombreRol;
+            await Rol.update(idRol, nuevoNombreRol);
 
-            // Asignar los nuevos privilegios seleccionados al rol
-            const privilegiosSeleccionados = [req.body.rol].flat();
-            const asignacionesPromises = privilegiosSeleccionados.map( async privilegio => {
-                console.log(privilegio)
-                const [idprivilegio] = await Rol.obtenerPrivilegioPorNombre(privilegio)
-                console.log(idprivilegio);
-                return Rol.asignarPrivilegio(idRol, idprivilegio[0].idprivilegio);
-            });
-            await Promise.all(asignacionesPromises);
+            // Verificar si hay privilegios seleccionados
+            if (req.body.rol) {
+                console.log([req.body.rol].flat())
+                // Eliminar todos los privilegios asociados al rol
+                await Rol.eliminarPrivilegios(idRol);
+
+                // Asignar los nuevos privilegios seleccionados al rol
+                const privilegiosSeleccionados = [req.body.rol].flat();
+                const asignacionesPromises = privilegiosSeleccionados.map( async privilegio => {
+                    console.log(privilegio)
+                    const [idprivilegio] = await Rol.obtenerPrivilegioPorNombre(privilegio)
+                    console.log(idprivilegio);
+                    return Rol.asignarPrivilegio(idRol, idprivilegio[0].idprivilegio);
+                });
+                await Promise.all(asignacionesPromises);
+            }
+
+            // Redirigir a la página de gestión de roles con un mensaje de éxito
+            req.session.mensaje = `El rol ${nuevoNombreRol} fue modificado correctamente`;
+            res.redirect('/gestionRoles');
         }
 
-        // Redirigir a la página de gestión de roles con un mensaje de éxito
-        req.session.mensaje = `El rol ${nuevoNombreRol} fue modificado correctamente`;
-        res.redirect('/gestionRoles');
     } catch (error) {
         console.error('Error al actualizar el rol:', error);
         req.session.error = 'Error al actualizar el rol';
